ipa: ipu3: Introduce IPAConfigInfo in IPC
IPAConfigInfo is a consolidated data structure passed from IPU3
pipeline-handler to IPU3 IPA. This provides a streamlined and
extensible way to provide the configuration data to IPA interface.

The IPA interface should be common to both closed-source and
open-source IPU3 IPAs. Hence, the structure encompasses additional
configuration information required to init and configure the
closed-source IPA as well.

Signed-off-by: Umang Jain <umang.jain@ideasonboard.com>
Change-Id: I174f2e0d0eee54a47477db37e99145aacaa6420b