ipa: ipu3: Rework IPAFrameContext
Currently, IPAFrameContext consolidates the values computed by the
active state of the algorithms, along with the values applied on
the sensor.

Moving ahead, we want to have a frame context associated with each
incoming request (or frame to be captured). This not necessarily should
be tied to the "active state" of the algorithms hence:
	- Rename current IPAFrameContext -> IPAActiveState
	  This will now reflect the latest active state of the
	  algorithms and has nothing to do with any frame-related
	  ops/values.
	- Re-instate IPAFrameContext with a sub-structure 'sensor'
	  currently storing the exposure and gain value.

Adapt the various access to the frame context to the new changes
as described above.

In subsequent patch, the re-instated IPAFrameContext will reflect a
frame number stored in it and will be used in conjunction with a queue.

Signed-off-by: Umang Jain <umang.jain@ideasonboard.com>
Change-Id: I373d7d509b4fb8687efaeb7875a6e3b96673e657