ipa: ipu3: Introduce a IPAFrameContext queue
Introduce a queue of IPAFrameContext which shall maintain the data
required to track each frame. Currently, it is storing only the
sensor controls values applied for the frame but new structures
can be introduced in IPAFrameContext to extend any frame-related
contexts (either in-processing or new frames being queued).

For example, this patch provides the foundation for the extension
of IPAFrameContext to store incoming request controls.

Signed-off-by: Umang Jain <umang.jain@ideasonboard.com>
Change-Id: I1122672497659e8af7525e5ce9ed3ba45cea872b