libcamera: v4l2_device: Adjust colorspace if pixel format is RGB
The V4L2_COLORSPACE_SRGB colorspace maps to ColorSpace::Srgb.
This is wrong as the V4L2_COLORSPACE_SRGB is ill-defined (defines
Rec 601 Y'CbCr encoding and limited range) in the kernel.

The RGB pixel formats should not use any Y'CbCr encoding and is always
full range. Adjust the colorspace before reporting back to the
userspace in such a situation.

Moving forwards, the ColorSpace::Srgb will be defined in the true sense
for RGB pixel formats.

Signed-off-by: Umang Jain <umang.jain@ideasonboard.com>